+++ packages/nimble-components/src/tabs-toolbar/styles.ts
@@ -27,4 +27,13 @@ export const styles = css`
             rgba(${borderRgbPartialColor}, 0.3);
         margin: ${smallPadding} ${mediumPadding};
     }
+
+    [part='start'] {
+        display: none;
+    }
+
+    [part='end'] {
+        margin-left: auto;
+        display: flex;
+    }
 `;