+++ packages/storybook/src/nimble/tabs/tabs-matrix.stories.ts
@@ -5,6 +5,10 @@ import { tabTag } from '@ni/nimble-components/dist/esm/tab';
 import { tabPanelTag } from '@ni/nimble-components/dist/esm/tab-panel';
 import { tabsToolbarTag } from '@ni/nimble-components/dist/esm/tabs-toolbar';
 import { tabsTag } from '@ni/nimble-components/dist/esm/tabs';
+import {
+    controlLabelFont,
+    controlLabelFontColor
+} from '@ni/nimble-components/dist/esm/theme-provider/design-tokens';
 import { createStory } from '../../utilities/storybook';
 import {
     createMatrixThemeStory,
@@ -16,7 +20,38 @@ import { hiddenWrapper } from '../../utilities/hidden';
 import { textCustomizationWrapper } from '../../utilities/text-customization';
 import { loremIpsum } from '../../utilities/lorem-ipsum';
 
-const tabsToolbarStates = [false, true] as const;
+const tabsToolbarStates = [
+    {
+        showToolbar: false,
+        showLeftButton: false,
+        showRightButtons: false,
+        label: 'No Toolbar'
+    },
+    {
+        showToolbar: true,
+        showLeftButton: false,
+        showRightButtons: false,
+        label: 'Toolbar with no buttons'
+    },
+    {
+        showToolbar: true,
+        showLeftButton: true,
+        showRightButtons: false,
+        label: 'Toolbar with left button'
+    },
+    {
+        showToolbar: true,
+        showLeftButton: false,
+        showRightButtons: true,
+        label: 'Toolbar with right buttons'
+    },
+    {
+        showToolbar: true,
+        showLeftButton: true,
+        showRightButtons: true,
+        label: 'Toolbar with left and right buttons'
+    }
+] as const;
 type TabsToolbarState = (typeof tabsToolbarStates)[number];
 
 const metadata: Meta = {
@@ -38,11 +73,19 @@ const component = (
     widthValue: WidthState
 
 ): ViewTemplate => html`
+    <label style="color: var(${controlLabelFontColor.cssCustomProperty}); font: var(${controlLabelFont.cssCustomProperty})">
+        ${toolbar.label} ${disabledName ? `(${disabledName})` : ''} ${widthValue ? `(${widthValue})` : ''}
+    </label>
     <${tabsTag} style="padding: 15px;${widthValue ? ` width: ${widthValue};` : ''}">
-        ${when(() => toolbar, html`
+        ${when(() => toolbar.showToolbar, html`
             <${tabsToolbarTag}>
-                <${buttonTag} appearance="ghost">Toolbar Button</${buttonTag}>
-                <${buttonTag} appearance="ghost" style="margin-left: auto;">Right-aligned Button</${buttonTag}>
+                ${when(() => toolbar.showLeftButton, html`
+                    <${buttonTag} appearance="ghost">Left Button</${buttonTag}>
+                `)}
+                ${when(() => toolbar.showRightButtons, html`
+                    <${buttonTag} appearance="ghost" slot="end">Right Button 1</${buttonTag}>
+                    <${buttonTag} appearance="ghost" slot="end">Right Button 2</${buttonTag}>
+                `)}
             </${tabsToolbarTag}>
         `)}
         <${tabTag}>Tab One</${tabTag}>
@@ -76,7 +119,7 @@ export const textCustomized: StoryFn = createMatrixThemeStory(
         html`
             <${tabsTag}>
                 Inner text
-                <${tabsToolbarTag}>Tabs toolbar</${tabsToolbarTag}>
+                <${tabsToolbarTag}><${buttonTag} appearance="ghost">Tabs toolbar</${buttonTag}></${tabsToolbarTag}>
                 <${tabTag}>Tab</${tabTag}>
             </${tabsTag}>
         `